requirements.txt
# Django Core
Django==5.0.2
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.10

# Authentication & Security
django-cors-headers==4.3.1
PyJWT==2.8.0

# API Documentation
drf-spectacular==0.27.1

# Data Processing & Export
django-filter==24.1
reportlab==4.2.5
Pillow==11.0.0

# Utilities
django-extensions==3.2.3
python-dotenv==1.0.1
whitenoise==6.6.0

# AI & Machine Learning
openai==1.58.1
numpy==2.2.0

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==5.1.0

# Development & Testing (optional)
# pytest==7.4.3
# pytest-django==4.7.0
# coverage==7.3.2
# black==23.11.0
# flake8==6.1.0
# isort==5.12.0
